ENH: Add Method for Removing Problematic Pixels
When building a mesh from an image, there are a number of configurations
of foreground/background pixels which require there to be two edges
connecting the same two vertices (creating a "hole" with zero area).
Unfortunately, the itk::QuadEdgeMesh class disallows this configuration,
resulting in missing faces when the itk::QuadEdgeMesh is used.

This patch adds a method, RemoveProblematicPixels, which identifies
the specific pixels which will cause non-manifold topology, and gives
the user the option to remove them prior to mesh generation.